backend/django_app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
django-filter==23.5
markdown==3.5
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
Pillow==10.1.0 